// admin_space/src/types/Employee.ts
// src/types/Employee.ts
export interface Employee {
    _id: string;
    name: string;
    email: string;
    dob: string;
    joiningDate: string;
    mobile: string;
    department: string;
    address: string;
    company: string;
    password: string;
    profilePicture: String;
    companyAddress: string;
    experience: string;
}